requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.3.4
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
random-username==1.0.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
